requirements.txt
playwright==1.42.0
python-dotenv==1.0.1
psycopg2-binary==2.9.9
sqlalchemy==2.0.27
pydantic==2.11.4
fastapi==0.110.0
uvicorn==0.27.1
streamlit==1.31.1
openai==1.12.0
qdrant-client==1.7.1
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.9 